requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
tqdm>=4.66.4
# To configure AWS upload bucket and prefix
python-dotenv>=1.0.0
PyPDF2>=3.0.1
# Dependency for AWS S3 uploads
boto3>=1.34.76
